refactor(cli): T7c slim run.rs dispatch arms — parse once, stage logic into modules (F33,F34)
