+++ crates/qedgen/src/main.rs
@@ -23,7 +23,7 @@ pub(crate) use adapt::{
     anchor_adapt, anchor_check, anchor_extractor, anchor_project, anchor_resolver,
     native_extractor, pinocchio_extractor, pinocchio_profile, program_model,
 };
-pub(crate) use cli::{AristotleCommands, Cli, Commands, RuntimeOverride, Target};
+pub(crate) use cli::{AristotleCommands, Cli, Commands, Target};
 pub(crate) use codegen::{
     asm2lean, banner, codegen_mir, codegen_shared, crucible_gen, fingerprint, integration_test,
     interface_gen, kani_impl, kani_mir, lean_gen_mir, lean_names, lean_sidecars, proptest_gen_mir,
